+++ ferricel-core/src/compiler/literals.rs
@@ -86,7 +86,7 @@ fn compile_string_literal(
             walrus::ir::StoreKind::I32_8 { atomic: false },
             walrus::ir::MemArg {
                 align: 1,
-                offset: offset as u32,
+                offset: offset as u64,
             },
         );
     }
@@ -137,7 +137,7 @@ fn compile_bytes_literal(
             walrus::ir::StoreKind::I32_8 { atomic: false },
             walrus::ir::MemArg {
                 align: 1,
-                offset: offset as u32,
+                offset: offset as u64,
             },
         );
     }
